feature: reservation 스케쥴링 기능

+++ src/main/java/org/hermez/paymenthistory/service/payapi/IamportService.java
@@ -49,7 +49,7 @@ public class IamportService {
    *
    * @param payResponse 결제 API( 아임포트 ) 서버에 요청한 응답
    * @param amount 클라이언트가 보낸 강의 가격
-   * @param courseId 클라이언트가 예약한 강의 키값
+   * @param merchantUid 클라이언트가 예약한 강의 키값
    * @throws IllegalArgumentException 클라이언트가 요청한 강의 가격이 결제 API( 아임포트 ) 서버와 hermez 서버에 저장된 가격과 다를 경우 발생
    */
   public void verifyPayment(IamportResponse<Payment> payResponse, int amount, String merchantUid) {
@@ -60,8 +60,10 @@ public class IamportService {
       log.info("[아임포트 서버(실결제 금액)과 강의 가격이 다릅니다] 아임포트 서버 금액 = {} , 클라이언트가 보낸 강의 가격 = {} ",iamportAmount ,amount);
       throw new IllegalStateException("클라이언트가 보낸 강의 가격이 아임포트 서버 결제 금액과 다릅니다.");
     }
-    Double payAmount = reservationRepository.findPayAmount(merchantUid);//강의에서 가져와야함
+    int payAmount = reservationRepository.findPayAmount(merchantUid);//강의에서 가져와야함
+
     //db에서 강의 가격과 결제금액 같은지 확인
+
     if (amount != payAmount) {
       log.info("[결제시 hermez 서버에 저장된 금액과 클라이언트가 보낸 강의 가격이 다릅니다] hermez 서버 금액 = {} , 클라이언트가 보낸 강의 가격 = {}",payAmount,amount);
       throw new IllegalStateException("결제시 hermez 서버에 저장된 금액과 클라이언트가 보낸 강의 가격이 다릅니다.");
